% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbStrDF.R
\name{dbStr}
\alias{dbStr}
\title{This function allows you to access Databases.}
\usage{
dbStr()
}
\arguments{
\item{db}{db}
}
\description{
This function allows you to access Databases.
}
\examples{
dbStr()
}
\keyword{db}
