% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnector.R
\name{mdbQuery}
\alias{mdbQuery}
\title{This function allows you to access Databases.}
\usage{
mdbQuery(query = NULL, username, password)
}
\arguments{
\item{query, }{username, password}
}
\description{
This function allows you to access Databases.
}
\examples{
mdbQuery()
}
\keyword{password}
\keyword{query,}
\keyword{username,}
