% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usepkg.R
\name{usepkg}
\alias{usepkg}
\title{This function allows you to access Databases.}
\usage{
usepkg(pkg)
}
\arguments{
\item{pkg, }{}
}
\description{
This function allows you to access Databases.
}
\examples{
usepkg()
}
\keyword{package,}
