% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_loop.R
\name{vif_loop}
\alias{vif_loop}
\title{This function allows you to access Databases.}
\usage{
vif_loop(in_frame, thresh = 10, trace = F, ...)
}
\arguments{
\item{pkg, }{}
}
\description{
This function allows you to access Databases.
}
\examples{
vif_loop()
}
\keyword{package,}
