% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbQuery.R
\name{dbQuery}
\alias{dbQuery}
\title{This function allows you to access Databases.}
\usage{
dbQuery(db = c(type = "dbname", driver = "jdbcDriverPath", url = "url", port =
  "port"), query = NULL, username, password)
}
\arguments{
\item{db(dbname, }{jdbcDriverPath, url, port), query, username, password}
}
\description{
This function allows you to access Databases.
}
\examples{
dbQuery()
}
\keyword{db(dbname,}
\keyword{jdbcDriverPath,}
\keyword{password}
\keyword{port),}
\keyword{query,}
\keyword{url,}
\keyword{username,}
